Fix naming consistency: renamed affiliates folder to affiliations to match component name
