+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Hero from "./_components/hero/hero";
 import Navbar from "./_components/navbar/navbar";
-import Affiliations from "./_components/affiliates/affiliations";
+import Affiliations from "./_components/affiliations/affiliations";
 export default function Home() {
   return (
     <div className="flex flex-col items-center">
